Исправлена задача в 7 практической работе

+++ students/21K0623/21K0623-p07/src/main/java/ru/mirea/practice/solutiontask1/Movable.java
@@ -1,11 +1,11 @@
 package ru.mirea.practice.solutiontask1;
 
 public interface Movable {
-    void moveUp(double yspeed);
+    void moveUp(int yspeed);
 
-    void moveDown(double yspeed);
+    void moveDown(int yspeed);
 
-    void moveLeft(double xspeed);
+    void moveLeft(int xspeed);
 
-    void moveRight(double xspeed);
+    void moveRight(int xspeed);
 }
