+++ students/21K0623/21K0623-p07/src/main/java/ru/mirea/practice/solutiontask1/MovablePoint.java
@@ -15,22 +15,22 @@ public class MovablePoint implements Movable {
     }
 
     @Override
-    public void moveUp(double yspeed) {
+    public void moveUp(int yspeed) {
         this.y += yspeed;
     }
 
     @Override
-    public void moveDown(double yspeed) {
+    public void moveDown(int yspeed) {
         this.y -= yspeed;
     }
 
     @Override
-    public void moveLeft(double xspeed) {
+    public void moveLeft(int xspeed) {
         this.x -= xspeed;
     }
 
     @Override
-    public void moveRight(double xspeed) {
+    public void moveRight(int xspeed) {
         this.x += xspeed;
     }
 }
